% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{available_instruments}
\alias{available_instruments}
\title{List available questionnaire instruments}
\usage{
available_instruments()
}
\value{
A character vector of instrument IDs supported by
\code{\link[=score_questionnaire]{score_questionnaire()}}.
}
\description{
List available questionnaire instruments
}
